// src/app/commonBussinessSRC/psComponents/ps-input-display-only/ps-display-only-currency-amount/ps-display-only-currency-amount.component.interfaces.ts
import { IOptionsPsBaseFieldExposed } from 'src/app/commonBussinessSRC/psServices/models/ps-common-bussiness-interfaces';
import { IOptionsPsActionImage, IOptionsPsInputAmount } from 'src/app/commonSRC/psServices/models/ps-common-interface';

import { IOptionsPsDropdownCurrenciesExposed } from '../../ps-select-dropdown/ps-dropdown-currencies/ps-dropdown-currencies.component.interfaces';

export interface IOptionsPsDisplayOnlyCurrencyAmountExposed extends IOptionsPsBaseFieldExposed {
    currenciesOptions?: IOptionsPsDropdownCurrenciesExposed;
    amountOptions?: IOptionsPsInputAmount;
    currency?: string;
    currencyCode?: number;
    currencyDisabled?: boolean;
    currencyFlagOptions?: IOptionsPsActionImage;
    amount?: number;
    currencyObj?: any;

}